requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
google-auth==1.31.0
gunicorn==20.1.0
idna==2.10
#mysqlclient==2.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
psycopg2-binary==2.9.1
django_extensions
pillow